+++ codegen/templates/ethereum/results/lib.rs
@@ -120,9 +120,7 @@ fn db_out(events: contract::Events) -> Result<DatabaseChanges, substreams::error
     // Loop over all the abis events to create changes
     events.approvals.into_iter().for_each(|evt| {
         tables
-            .create_row("approval", format!("{}-{}", evt.evt_tx_hash, evt.evt_index))
-            .set("evt_tx_hash", evt.evt_tx_hash)
-            .set("evt_index", evt.evt_index)
+            .create_row("approval", [("evt_tx_hash", evt.evt_tx_hash),("evt_index", evt.evt_index.to_string())])
             .set("evt_block_time", evt.evt_block_time.unwrap())
             .set("evt_block_number", evt.evt_block_number)
             .set("approved", Hex(&evt.approved).to_string())
@@ -131,9 +129,7 @@ fn db_out(events: contract::Events) -> Result<DatabaseChanges, substreams::error
     });
     events.approval_for_alls.into_iter().for_each(|evt| {
         tables
-            .create_row("approval_for_all", format!("{}-{}", evt.evt_tx_hash, evt.evt_index))
-            .set("evt_tx_hash", evt.evt_tx_hash)
-            .set("evt_index", evt.evt_index)
+            .create_row("approval_for_all", [("evt_tx_hash", evt.evt_tx_hash),("evt_index", evt.evt_index.to_string())])
             .set("evt_block_time", evt.evt_block_time.unwrap())
             .set("evt_block_number", evt.evt_block_number)
             .set("approved", evt.approved)
@@ -142,9 +138,7 @@ fn db_out(events: contract::Events) -> Result<DatabaseChanges, substreams::error
     });
     events.ownership_transferreds.into_iter().for_each(|evt| {
         tables
-            .create_row("ownership_transferred", format!("{}-{}", evt.evt_tx_hash, evt.evt_index))
-            .set("evt_tx_hash", evt.evt_tx_hash)
-            .set("evt_index", evt.evt_index)
+            .create_row("ownership_transferred", [("evt_tx_hash", evt.evt_tx_hash),("evt_index", evt.evt_index.to_string())])
             .set("evt_block_time", evt.evt_block_time.unwrap())
             .set("evt_block_number", evt.evt_block_number)
             .set("new_owner", Hex(&evt.new_owner).to_string())
@@ -152,9 +146,7 @@ fn db_out(events: contract::Events) -> Result<DatabaseChanges, substreams::error
     });
     events.transfers.into_iter().for_each(|evt| {
         tables
-            .create_row("transfer", format!("{}-{}", evt.evt_tx_hash, evt.evt_index))
-            .set("evt_tx_hash", evt.evt_tx_hash)
-            .set("evt_index", evt.evt_index)
+            .create_row("transfer", [("evt_tx_hash", evt.evt_tx_hash),("evt_index", evt.evt_index.to_string())])
             .set("evt_block_time", evt.evt_block_time.unwrap())
             .set("evt_block_number", evt.evt_block_number)
             .set("from", Hex(&evt.from).to_string())