+++ codegen/templates/ethereum/results/multiple_contracts/schema.sql
@@ -5,7 +5,8 @@ CREATE TABLE IF NOT EXISTS moonbird_approval (
     "evt_block_number" DECIMAL,
     "approved" VARCHAR(40),
     "owner" VARCHAR(40),
-    "token_id" DECIMAL
+    "token_id" DECIMAL,
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
 CREATE TABLE IF NOT EXISTS moonbird_approval_for_all (
     "evt_tx_hash" VARCHAR(64),
@@ -14,21 +15,24 @@ CREATE TABLE IF NOT EXISTS moonbird_approval_for_all (
     "evt_block_number" DECIMAL,
     "approved" BOOL,
     "operator" VARCHAR(40),
-    "owner" VARCHAR(40)
+    "owner" VARCHAR(40),
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
 CREATE TABLE IF NOT EXISTS moonbird_expelled (
     "evt_tx_hash" VARCHAR(64),
     "evt_index" INT,
     "evt_block_time" TIMESTAMP,
     "evt_block_number" DECIMAL,
-    "token_id" DECIMAL
+    "token_id" DECIMAL,
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
 CREATE TABLE IF NOT EXISTS moonbird_nested (
     "evt_tx_hash" VARCHAR(64),
     "evt_index" INT,
     "evt_block_time" TIMESTAMP,
     "evt_block_number" DECIMAL,
-    "token_id" DECIMAL
+    "token_id" DECIMAL,
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
 CREATE TABLE IF NOT EXISTS moonbird_ownership_transferred (
     "evt_tx_hash" VARCHAR(64),
@@ -36,14 +40,16 @@ CREATE TABLE IF NOT EXISTS moonbird_ownership_transferred (
     "evt_block_time" TIMESTAMP,
     "evt_block_number" DECIMAL,
     "new_owner" VARCHAR(40),
-    "previous_owner" VARCHAR(40)
+    "previous_owner" VARCHAR(40),
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
 CREATE TABLE IF NOT EXISTS moonbird_paused (
     "evt_tx_hash" VARCHAR(64),
     "evt_index" INT,
     "evt_block_time" TIMESTAMP,
     "evt_block_number" DECIMAL,
-    "account" VARCHAR(40)
+    "account" VARCHAR(40),
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
 CREATE TABLE IF NOT EXISTS moonbird_refund (
     "evt_tx_hash" VARCHAR(64),
@@ -51,7 +57,8 @@ CREATE TABLE IF NOT EXISTS moonbird_refund (
     "evt_block_time" TIMESTAMP,
     "evt_block_number" DECIMAL,
     "amount" DECIMAL,
-    "buyer" VARCHAR(40)
+    "buyer" VARCHAR(40),
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
 CREATE TABLE IF NOT EXISTS moonbird_revenue (
     "evt_tx_hash" VARCHAR(64),
@@ -60,7 +67,8 @@ CREATE TABLE IF NOT EXISTS moonbird_revenue (
     "evt_block_number" DECIMAL,
     "amount" DECIMAL,
     "beneficiary" VARCHAR(40),
-    "num_purchased" DECIMAL
+    "num_purchased" DECIMAL,
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
 CREATE TABLE IF NOT EXISTS moonbird_role_admin_changed (
     "evt_tx_hash" VARCHAR(64),
@@ -69,7 +77,8 @@ CREATE TABLE IF NOT EXISTS moonbird_role_admin_changed (
     "evt_block_number" DECIMAL,
     "new_admin_role" TEXT,
     "previous_admin_role" TEXT,
-    "role" TEXT
+    "role" TEXT,
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
 CREATE TABLE IF NOT EXISTS moonbird_role_granted (
     "evt_tx_hash" VARCHAR(64),
@@ -78,7 +87,8 @@ CREATE TABLE IF NOT EXISTS moonbird_role_granted (
     "evt_block_number" DECIMAL,
     "account" VARCHAR(40),
     "role" TEXT,
-    "sender" VARCHAR(40)
+    "sender" VARCHAR(40),
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
 CREATE TABLE IF NOT EXISTS moonbird_role_revoked (
     "evt_tx_hash" VARCHAR(64),
@@ -87,7 +97,8 @@ CREATE TABLE IF NOT EXISTS moonbird_role_revoked (
     "evt_block_number" DECIMAL,
     "account" VARCHAR(40),
     "role" TEXT,
-    "sender" VARCHAR(40)
+    "sender" VARCHAR(40),
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
 CREATE TABLE IF NOT EXISTS moonbird_transfer (
     "evt_tx_hash" VARCHAR(64),
@@ -96,21 +107,24 @@ CREATE TABLE IF NOT EXISTS moonbird_transfer (
     "evt_block_number" DECIMAL,
     "from" VARCHAR(40),
     "to" VARCHAR(40),
-    "token_id" DECIMAL
+    "token_id" DECIMAL,
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
 CREATE TABLE IF NOT EXISTS moonbird_unnested (
     "evt_tx_hash" VARCHAR(64),
     "evt_index" INT,
     "evt_block_time" TIMESTAMP,
     "evt_block_number" DECIMAL,
-    "token_id" DECIMAL
+    "token_id" DECIMAL,
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
 CREATE TABLE IF NOT EXISTS moonbird_unpaused (
     "evt_tx_hash" VARCHAR(64),
     "evt_index" INT,
     "evt_block_time" TIMESTAMP,
     "evt_block_number" DECIMAL,
-    "account" VARCHAR(40)
+    "account" VARCHAR(40),
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
 CREATE TABLE IF NOT EXISTS bayc_approval (
     "evt_tx_hash" VARCHAR(64),
@@ -119,7 +133,8 @@ CREATE TABLE IF NOT EXISTS bayc_approval (
     "evt_block_number" DECIMAL,
     "approved" VARCHAR(40),
     "owner" VARCHAR(40),
-    "token_id" DECIMAL
+    "token_id" DECIMAL,
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
 CREATE TABLE IF NOT EXISTS bayc_approval_for_all (
     "evt_tx_hash" VARCHAR(64),
@@ -128,7 +143,8 @@ CREATE TABLE IF NOT EXISTS bayc_approval_for_all (
     "evt_block_number" DECIMAL,
     "approved" BOOL,
     "operator" VARCHAR(40),
-    "owner" VARCHAR(40)
+    "owner" VARCHAR(40),
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
 CREATE TABLE IF NOT EXISTS bayc_ownership_transferred (
     "evt_tx_hash" VARCHAR(64),
@@ -136,7 +152,8 @@ CREATE TABLE IF NOT EXISTS bayc_ownership_transferred (
     "evt_block_time" TIMESTAMP,
     "evt_block_number" DECIMAL,
     "new_owner" VARCHAR(40),
-    "previous_owner" VARCHAR(40)
+    "previous_owner" VARCHAR(40),
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
 CREATE TABLE IF NOT EXISTS bayc_transfer (
     "evt_tx_hash" VARCHAR(64),
@@ -145,5 +162,7 @@ CREATE TABLE IF NOT EXISTS bayc_transfer (
     "evt_block_number" DECIMAL,
     "from" VARCHAR(40),
     "to" VARCHAR(40),
-    "token_id" DECIMAL
+    "token_id" DECIMAL,
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
+