+++ codegen/templates/ethereum/schema.sql
@@ -5,7 +5,8 @@ CREATE TABLE IF NOT EXISTS approval (
     "evt_block_number" DECIMAL,
     "approved" VARCHAR(40),
     "owner" VARCHAR(40),
-    "token_id" DECIMAL
+    "token_id" DECIMAL,
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
 CREATE TABLE IF NOT EXISTS approval_for_all (
     "evt_tx_hash" VARCHAR(64),
@@ -14,7 +15,8 @@ CREATE TABLE IF NOT EXISTS approval_for_all (
     "evt_block_number" DECIMAL,
     "approved" BOOL,
     "operator" VARCHAR(40),
-    "owner" VARCHAR(40)
+    "owner" VARCHAR(40),
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
 CREATE TABLE IF NOT EXISTS ownership_transferred (
     "evt_tx_hash" VARCHAR(64),
@@ -22,7 +24,8 @@ CREATE TABLE IF NOT EXISTS ownership_transferred (
     "evt_block_time" TIMESTAMP,
     "evt_block_number" DECIMAL,
     "new_owner" VARCHAR(40),
-    "previous_owner" VARCHAR(40)
+    "previous_owner" VARCHAR(40),
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
 CREATE TABLE IF NOT EXISTS transfer (
     "evt_tx_hash" VARCHAR(64),
@@ -31,5 +34,7 @@ CREATE TABLE IF NOT EXISTS transfer (
     "evt_block_number" DECIMAL,
     "from" VARCHAR(40),
     "to" VARCHAR(40),
-    "token_id" DECIMAL
+    "token_id" DECIMAL,
+    PRIMARY KEY(evt_tx_hash,evt_index)
 );
+